+++ src/runtime/locale/fr.ts
@@ -1,20 +1,20 @@
 import { LocaleStrings } from '../types'
 
 export default {
+  accept: 'Accepter',
   acceptAll: 'Tout accepter',
-  acceptNecessary: "J'accepte le nécessaire",
-  barDescription:
+  bannerDescription:
     'Nous utilisons des cookies d’origine et des cookies tiers. Ces cookies sont destinés à vous offrir une navigation optimisée sur ce site web et de nous donner un aperçu de son utilisation, en vue de l’amélioration des services que nous offrons. En poursuivant votre navigation, nous considérons que vous acceptez l’usage des cookies.',
-  barTitle: 'Cookies',
-  blockedIframe: 'Pour voir cela, veuillez activer les cookies fonctionnels',
+  bannerTitle: 'Cookies',
   close: 'Fermer',
+  cookiesFunctional: 'Cookies fonctionnels',
+  cookiesNecessary: 'Les cookies obligatoires',
+  cookiesOptional: 'Les cookies optionnels',
+  decline: "J'accepte le nécessaire",
   declineAll: 'Tout refuser',
-  functional: 'Cookies fonctionnels',
   here: 'ici',
+  iframeBlocked: 'Pour voir cela, veuillez activer les cookies fonctionnels',
   manageCookies: 'Gérer les cookies',
-  necessary: 'Les cookies obligatoires',
-  none: 'rien',
-  optional: 'Les cookies optionnels',
   save: 'Sauvegarder',
-  unsaved: 'Vous avez des paramètres non sauvegardés',
+  settingsUnsaved: 'Vous avez des paramètres non sauvegardés',
 } as LocaleStrings