+++ src/runtime/locale/nl.ts
@@ -1,21 +1,21 @@
 import { LocaleStrings } from '../types'
 
 export default {
+  accept: 'Aanvaarden',
   acceptAll: 'Accepteer alles',
-  acceptNecessary: 'Accepteer noodzakelijk',
-  barDescription:
+  bannerDescription:
     'We gebruiken onze eigen cookies en third-party cookies om onze site te tonen en om te leren hoe deze gebruikt wordt, met het oog om de services die we verlenen te verbeteren. Door verder te gaan op onze site gaan we ervanuit dat hiermee akkoord gegaan wordt.',
-  barTitle: 'Cookies',
-  blockedIframe:
-    'Om dit te kunnen bekijken dienen functionele cookies geaccepteerd te worden',
+  bannerTitle: 'Cookies',
   close: 'Sluiten',
+  cookiesFunctional: 'Functionele cookies',
+  cookiesNecessary: 'Noodzakelijke cookies',
+  cookiesOptional: 'Optionele cookies',
+  decline: 'Accepteer noodzakelijk',
   declineAll: 'Verwijder alles',
-  functional: 'Functionele cookies',
   here: 'hier',
+  iframeBlocked:
+    'Om dit te kunnen bekijken dienen functionele cookies geaccepteerd te worden',
   manageCookies: 'Beheer cookies',
-  necessary: 'Noodzakelijke cookies',
-  none: 'geen',
-  optional: 'Optionele cookies',
   save: 'Opslaan',
-  unsaved: 'Er zijn niet-opgeslagen instellingen',
+  settingsUnsaved: 'Er zijn niet-opgeslagen instellingen',
 } as LocaleStrings