+++ src/runtime/locale/pt.ts
@@ -1,21 +1,21 @@
 import { LocaleStrings } from '../types'
 
 export default {
-  acceptAll: 'Aceitar todos',
-  acceptNecessary: 'Eu aceito o necessário',
-  barDescription:
+  accept: 'Aceitar todos',
+  acceptAll: 'Aceitar',
+  bannerDescription:
     'Nós utilizamos os nossos próprios cookies e de terceiros para que possamos lhe mostrar este site e compreender a forma como o utiliza, de forma a melhorarmos os serviços que oferecemos. Ao continuar a navegar no site, consideramos que aceitou a utilização de cookies.',
-  barTitle: 'Cookies',
-  blockedIframe:
-    'Para visualizar isto, por favor, active os cookies funcionais',
+  bannerTitle: 'Cookies',
   close: 'Fechar',
+  cookiesFunctional: 'Cookies funcionais',
+  cookiesNecessary: 'Cookies necessários',
+  cookiesOptional: 'Cookies opcionais',
+  decline: 'Eu aceito o necessário',
   declineAll: 'Apagar todos',
-  functional: 'Cookies funcionais',
   here: 'aqui',
+  iframeBlocked:
+    'Para visualizar isto, por favor, active os cookies funcionais',
   manageCookies: 'Gerir cookies',
-  necessary: 'Cookies necessários',
-  none: 'Nenhum',
-  optional: 'Cookies opcionais',
   save: 'Gravar',
-  unsaved: 'Tem alterações não guardadas',
+  settingsUnsaved: 'Tem alterações não guardadas',
 } as LocaleStrings