+++ src/runtime/types.ts
@@ -36,47 +36,47 @@ export interface Cookie {
 }
 
 export interface LocaleStrings {
+  accept: string
   acceptAll: string
-  acceptNecessary: string
-  barDescription: string
-  barTitle: string
-  blockedIframe: string
+  bannerDescription: string
+  bannerTitle: string
   close: string
+  cookiesFunctional: string
+  cookiesNecessary: string
+  cookiesOptional: string
+  iframeBlocked: string
+  decline: string
   declineAll: string
-  functional: string
   here: string
   manageCookies: string
-  necessary: string
-  none: string
-  optional: string
   save: string
-  unsaved: string
+  settingsUnsaved: string
 }
 
 export interface ModuleOptions {
-  barPosition?:
+  barPosition:
     | 'top-left'
     | 'top-right'
     | 'top-full'
     | 'bottom-left'
     | 'bottom-right'
     | 'bottom-full'
-  colors?: false | Record<string, any>
+  colors: false | Record<string, any>
+  cookieExpiryOffsetMs: number
+  cookieNameCookiesEnabledIds: string
+  cookieNameIsConsentGiven: string
   cookies: {
     necessary: Cookie[]
     optional: Cookie[]
   }
-  cookieExpiryOffsetMs: number
-  cookieNameIsConsentGiven: string
-  cookieNameCookiesEnabledIds: string
-  isAcceptNecessaryButtonEnabled?: boolean
-  isControlButtonEnabled?: boolean
-  isCookieIdVisible?: boolean
-  isCssEnabled?: boolean
-  isCssPolyfillEnabled?: boolean
-  isDashInDescriptionEnabled?: boolean
-  isIframeBlocked?: boolean | { initialState: boolean }
-  domain?: string
+  domain: string
+  isAcceptNecessaryButtonEnabled: boolean
+  isControlButtonEnabled: boolean
+  isCookieIdVisible: boolean
+  isCssEnabled: boolean
+  isCssPolyfillEnabled: boolean
+  isDashInDescriptionEnabled: boolean
+  isIframeBlocked: boolean | { initialState: boolean }
   locales: Locale[]
   localeTexts: PartialRecord<Locale, LocaleStrings>
 }